add thumnail create params

+++ src/main/java/com/hig/hwangingyu/utils/FfmpegProcess.java
@@ -11,35 +11,38 @@ public class FfmpegProcess {
  // 
     private static final String[] argvs = { 
         // webm으로 바꿔보기
+        "-flags", "+global_header", "-r", "30",
         "-pix_fmt", "yuv420p",
         "-c:v", "h264", "-b:v", "9000K", "-bufsize:v:0", "9000K/2",
-        "-g:v", "60", "-keyint_min:v", "60", "-sc_threshold:v", "0",
+        "-g:v", "30", "-keyint_min:v", "30", "-sc_threshold:v", "0",
         "-color_primaries", "bt709", "-color_trc", "bt709", "-colorspace", "bt709",
         "-c:a", "aac", "-ar", "44100", "-b:a", "96k", // audio codec config
         "-map", "0:v:0",
-        "-map", "0:a:0",
-        "-adaptation_sets", "id=0,seg_duration=1.001,streams=v id=1,seg_duration=1.001,streams=a",
-        "-use_timeline", "1",
+        "-map", "0:a:0?",
+        "-preset", "ultrafast",
+        //"-tune", "zerolatency",
+        "-adaptation_sets", "id=0,seg_duration=1.000,streams=v id=1,seg_duration=1.000,streams=a",
+        "-use_timeline", "0",
         "-streaming", "1",
-        "-window_size", "5",
+        "-window_size", "3",
         "-frag_type", "every_frame",
         "-ldash", "1",
         "-utc_timing_url", "https://time.akamai.com?iso&amp;ms",
         "-format_options", "movflags=cmaf",
-        "-timeout", "0.1",
+        "-timeout", "0.5",
         "-write_prft", "1",
-        "-target_latency", "4.0",
+        "-target_latency", "3.0",
         "-http_user_agent", "Akamai_Broadcaster_v1.0",
         "-http_persistent", "1",
         "-media_seg_name", "chunk-stream_$RepresentationID$-$Number%05d$.$ext$",
         "-init_seg_name", "init-stream_$RepresentationID$.$ext$",
-        "-update_period", "1",
-        "-f", "dash"
+        "-f", "dash",
+        
     };
     private static final String[] listenOption = {"-listen", "1", "-i"};
     private static final String[] basicOption = {"-i", "-"};
     private final String streamerPath = "/home/hig/jetson-ffmpeg/build/ffmpeg/ffmpeg";
-    private final String filePath = getClass().getClassLoader().getResource(".").getFile() + "static/stream/";
+    private final String filePath = getClass().getClassLoader().getResource(".").getFile() + "static/stream/vid/";
     private Process ffmpegProcess;
     private OutputStream ffmpegInput;
     private File log;
@@ -69,7 +72,14 @@ public class FfmpegProcess {
         for (String i : argvs) {
             commands.add(i);
         }
+        
         commands.add(streamFolder + "/test.mpd");
+        String[] tcmd = {"-vf","fps=1/10", "-y", "-q:v", "1", "-update", "1"};
+        for(String i : tcmd) {
+            commands.add(i);
+        }
+        commands.add(streamFolder+"/thumbnail.jpg");
+        
 
         ProcessBuilder ffmpegProcessBuilder = new ProcessBuilder(commands);
         ffmpegProcessBuilder.redirectErrorStream(true);
@@ -82,15 +92,15 @@ public class FfmpegProcess {
     }
     public boolean destroy() {
         try {
-            while(ffmpegProcess.isAlive()) {
-                ffmpegProcess.destroyForcibly();
+            if(ffmpegProcess.isAlive()) {
+                ffmpegProcess.destroy();
             }
             
-            while(log.exists()) {
+            if(log.exists()) {
                 log.delete();
             }
 
-            while(streamFolder.exists()) {
+            if(streamFolder.exists()) {
                 deleteDir(streamFolder);
             }
